Pull utilities out to separate file
