+++ src/Models/Context/OatPageContext/OatPageContextUtils.ts
@@ -1,0 +1,148 @@
+import { ProjectData } from '../../../Pages/OATEditorPage/Internal/Classes';
+import { IOATFile } from '../../../Pages/OATEditorPage/Internal/Classes/OatTypes';
+import {
+    convertDtdlInterfacesToModels,
+    storeLastUsedProjectId,
+    storeOntologiesToStorage
+} from '../../Services/OatUtils';
+import { createGUID, deepCopy } from '../../Services/Utils';
+import { isOatContextStorageEnabled, logDebugConsole } from './OatPageContext';
+import { IOatPageContextState } from './OatPageContext.types';
+
+export function createProject(
+    name: string,
+    namespace: string,
+    draft: IOatPageContextState
+) {
+    const id = createGUID();
+    const project = new ProjectData(
+        name,
+        namespace.replace(/ /g, ''),
+        [],
+        [],
+        [],
+        []
+    );
+    draft.ontologyFiles.push({ id: id, data: project });
+
+    saveData(draft);
+    switchCurrentProject(id, draft);
+
+    logDebugConsole(
+        'debug',
+        `Created new project with id: ${id}, {project}`,
+        project
+    );
+}
+
+export function switchCurrentProject(
+    projectId: string,
+    draft: IOatPageContextState
+) {
+    // do the swap
+    draft.currentOntologyId = projectId;
+    saveLastProjectId(draft.currentOntologyId);
+    const selectedFile = draft.ontologyFiles.find(
+        (x) => x.id === draft.currentOntologyId
+    );
+    if (selectedFile) {
+        const data = selectedFile.data;
+        const projectToOpen = new ProjectData(
+            data.projectName,
+            data.namespace,
+            convertDtdlInterfacesToModels(data.models),
+            data.modelPositions,
+            data.modelsMetadata,
+            data.templates
+        );
+        mapProjectOntoState(draft, projectToOpen);
+        logDebugConsole(
+            'debug',
+            `Switched to project: ${draft.currentOntologyProjectName} (${draft.currentOntologyId}). {project}`,
+            projectToOpen
+        );
+    } else {
+        logDebugConsole(
+            'warn',
+            `Project not found in storage. Unable to find the current project to ${draft.currentOntologyId}`
+        );
+    }
+}
+
+/** TODO: remove this helper when we move the project data into a sub object on the state */
+export function convertStateToProject(
+    draft: IOatPageContextState
+): ProjectData {
+    // NOTE: need to recreate the arrays to break proxy links that were causing issues downstream
+    const project = new ProjectData(
+        draft.currentOntologyProjectName || '',
+        draft.currentOntologyNamespace,
+        convertDtdlInterfacesToModels(draft.currentOntologyModels),
+        Array.from(draft.currentOntologyModelPositions),
+        Array.from(draft.currentOntologyModelMetadata),
+        Array.from(draft.currentOntologyTemplates)
+    );
+    // console.log('***Converted project', project, current(draft));
+
+    return project;
+}
+
+export function mapProjectOntoState(
+    draft: IOatPageContextState,
+    projectToOpen: ProjectData
+) {
+    draft.currentOntologyModelPositions = projectToOpen.modelPositions;
+    draft.currentOntologyModels = projectToOpen.models;
+    draft.currentOntologyModelMetadata = projectToOpen.modelsMetadata;
+    draft.currentOntologyNamespace = projectToOpen.namespace;
+    draft.currentOntologyProjectName = projectToOpen.projectName || '';
+    draft.currentOntologyTemplates = projectToOpen.templates;
+}
+
+/** saves all the data to local storage */
+export function saveData(draft: IOatPageContextState): void {
+    const selectedOntology = draft.ontologyFiles.find(
+        (x) => x.id === draft.currentOntologyId
+    );
+    if (selectedOntology) {
+        selectedOntology.data = convertStateToProject(draft); // update the files list with the latest data
+        const localDraft = deepCopy(draft); // copy to avoid the delayed proxy references
+        saveOntologyFiles(localDraft.ontologyFiles);
+        saveLastProjectId(localDraft.currentOntologyId);
+    } else {
+        logDebugConsole(
+            'warn',
+            `Unable to persist the state data to local storage. Onotology with id: ${draft.currentOntologyId} wasn't found in storage.`
+        );
+    }
+}
+
+export function saveLastProjectId(projectId: string): void {
+    logDebugConsole(
+        'debug',
+        'Saving current project id to storage.',
+        projectId
+    );
+    isOatContextStorageEnabled && storeLastUsedProjectId(projectId);
+}
+
+/**
+ * Writes the collection of files to storage
+ * @param files all files to be stored
+ */
+export function saveOntologyFiles(files: IOATFile[]): void {
+    if (isOatContextStorageEnabled) {
+        storeOntologiesToStorage(files);
+        logDebugConsole(
+            'debug',
+            `Saved ${files.length} files to storage. {files}`,
+            files
+        );
+    } else {
+        logDebugConsole(
+            'warn',
+            'Storage disabled. Skipping saving files. {files}',
+            files
+        );
+    }
+}
